move Round serde into object
